backend/requirements.txt
Django>=4.0
psycopg2-binary
djangorestframework
django-cors-headers